requirements.txt
# Core dependencies - all local, no APIs
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
bitsandbytes>=0.41.0
gradio>=4.0.0
ollama>=0.1.0
PyPDF2>=3.0.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
psutil>=5.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
